+++ xenon-common/src/main/java/com/vmware/xenon/services/common/MigrationTaskService.java
@@ -32,6 +32,7 @@ import java.util.TreeSet;
 import java.util.concurrent.CompletionException;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.ConcurrentSkipListSet;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
@@ -287,6 +288,28 @@ public class MigrationTaskService extends StatefulService {
          * finished successfully.
          */
         public Long latestSourceUpdateTimeMicros = 0L;
+
+        /**
+         * Child options used by the factory being migrated.
+         */
+        public EnumSet<ServiceOption> factoryChildOptions;
+
+        /**
+         * Contains self-links that were selected for migration.
+         *
+         * This property is not indexed and is only used for book-keeping during
+         * the migration process.
+         */
+        public Set<String> migratedSelfLinks = new ConcurrentSkipListSet<>();
+
+        /**
+         * Contains self-links that were not-selected for migration because
+         * of documentOwner mismatches.
+         *
+         * This property is not indexed and is only used for book-keeping during
+         * the migration process.
+         */
+        public Set<String> nonMigratedSelfLinks = new ConcurrentSkipListSet<>();
 
         @Override
         public String toString() {
@@ -656,6 +679,7 @@ public class MigrationTaskService extends StatefulService {
         this.sendWithDeferredResult(configGet)
                 .thenCompose(op -> {
                     FactoryServiceConfiguration factoryConfig = op.getBody(FactoryServiceConfiguration.class);
+                    currentState.factoryChildOptions = factoryConfig.childOptions;
 
                     QueryTask countQuery = QueryTask.Builder.createDirectTask()
                             .addOption(QueryOption.COUNT)
@@ -756,6 +780,22 @@ public class MigrationTaskService extends StatefulService {
         // will call here with empty currentPageLinks.
         // In that case, this has processed all entries, thus self patch to mark finish, then exit.
         if (currentPageLinks.isEmpty()) {
+            if (!currentState.nonMigratedSelfLinks.isEmpty()) {
+                logSevere("Some documents were not migrated due to ownership mismatches. SelfLinks: %s",
+                        Utils.toJson(currentState.nonMigratedSelfLinks));
+
+                failTask(new IllegalStateException(
+                        String.format("%d documents were not migrated due to ownership mismatches.",
+                                currentState.nonMigratedSelfLinks.size())));
+
+                // These fields are only used for bookkeeping. They don't need to be
+                // persisted, hence setting them to NULL so that they get garbage collected.
+                currentState.migratedSelfLinks = null;
+                currentState.nonMigratedSelfLinks = null;
+
+                return;
+            }
+
             patchToFinished(lastUpdateTimesPerOwner);
             return;
         }
@@ -823,8 +863,20 @@ public class MigrationTaskService extends StatefulService {
 
                             URI hostUri = getHostUri(op);
                             hostUriByResult.put(doc, hostUri);
+
+                            if (!currentState.factoryChildOptions.contains(ServiceOption.ON_DEMAND_LOAD)) {
+                                // save selfLinks that were selected for migration.
+                                currentState.nonMigratedSelfLinks.remove(document.documentSelfLink);
+                                currentState.migratedSelfLinks.add(document.documentSelfLink);
+                            }
                         } else {
                             ownerMissMatched++;
+
+                            // save selfLinks that were not selected due to own mismatch.
+                            if (!currentState.factoryChildOptions.contains(ServiceOption.ON_DEMAND_LOAD) &&
+                                    !currentState.migratedSelfLinks.contains(document.documentSelfLink)) {
+                                currentState.nonMigratedSelfLinks.add(document.documentSelfLink);
+                            }
                         }
                     }