+++ xenon-common/src/test/java/com/vmware/xenon/services/common/TestMigrationTaskService.java
@@ -32,6 +32,7 @@ import java.util.Collections;
 import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -647,17 +648,11 @@ public class TestMigrationTaskService extends BasicReusableHostTestCase {
     @Test
     public void successMigrateOnlyDocumentsUpdatedAfterTime() throws Throwable {
         // create object in host
-        List<ExampleServiceState> states = createExampleDocuments(this.exampleSourceFactory, getSourceHost(),
-                this.serviceCount);
-        List<URI> uris = getFullUri(getSourceHost(), states);
+        createExampleDocuments(this.exampleSourceFactory, getSourceHost(), this.serviceCount);
 
-        long time = getSourceHost()
-                .getServiceState(EnumSet.noneOf(TestProperty.class), ExampleServiceState.class, uris)
-                .values()
-                .stream()
-                .mapToLong(d -> d.documentUpdateTimeMicros)
-                .max().orElse(0);
-        assertTrue("max upateTime should not be 0", time > 0);
+        // Save the current time after the first batch has been created and
+        // before moving towards the second batch.
+        long time = Utils.getNowMicrosUtc();
 
         List<ExampleServiceState> newStates = createExampleDocuments(this.exampleSourceFactory, getSourceHost(),
                 this.serviceCount, false);
@@ -757,6 +752,68 @@ public class TestMigrationTaskService extends BasicReusableHostTestCase {
 
         getDestinationHost().getServiceState(EnumSet.noneOf(TestProperty.class),
                 ExampleServiceState.class, uris);
+    }
+
+    @Test
+    public void failMigrationWithDocumentOwnerMismatch() throws Throwable {
+        // create objects in the source node-group
+        createExampleDocuments(this.exampleSourceFactory, getSourceHost(), this.serviceCount);
+
+        try {
+            // disable peer synchronization on each source host.
+            Iterator<VerificationHost> peerIt = this.host.getInProcessHostMap().values().iterator();
+            VerificationHost peerNode = null;
+            while (peerIt.hasNext()) {
+                peerNode = peerIt.next();
+                peerNode.setPeerSynchronizationEnabled(false);
+            }
+            assertNotNull(peerNode);
+
+            // Stop the last peerNode and add a new node to the group
+            this.host.stopHost(peerNode);
+            VerificationHost newHost = VerificationHost.create(0);
+            newHost.setMaintenanceIntervalMicros(TimeUnit.MILLISECONDS.toMicros(
+                    VerificationHost.FAST_MAINT_INTERVAL_MILLIS));
+            newHost.setPeerSynchronizationEnabled(false);
+            newHost.start();
+            newHost.waitForServiceAvailable(ExampleService.FACTORY_LINK);
+            this.host.addPeerNode(newHost);
+            this.host.joinNodesAndVerifyConvergence(this.nodeCount, true);
+
+            // Kick-off migration
+            MigrationTaskService.State migrationState = validMigrationState(
+                    ExampleService.FACTORY_LINK);
+            TestContext ctx = this.host.testCreate(1);
+            String[] out = new String[1];
+            Operation op = Operation.createPost(this.destinationFactoryUri)
+                    .setBody(migrationState)
+                    .setReferer(this.host.getUri())
+                    .setCompletion((o, e) -> {
+                        if (e != null) {
+                            this.host.log("Post service error: %s", Utils.toString(e));
+                            ctx.failIteration(e);
+                            return;
+                        }
+                        out[0] = o.getBody(State.class).documentSelfLink;
+                        ctx.completeIteration();
+                    });
+            getDestinationHost().send(op);
+            ctx.await();
+
+            // Wait for the migration task to fail
+            State waitForServiceCompletion = waitForServiceCompletion(out[0], getDestinationHost());
+            assertEquals(waitForServiceCompletion.taskInfo.stage, TaskStage.FAILED);
+
+            ServiceStats stats = getStats(out[0], getDestinationHost());
+            long ownerMismatchedDocuments = (long)stats.entries.get(MigrationTaskService.STAT_NAME_OWNER_MISMATCH_COUNT).latestValue;
+            assertTrue(ownerMismatchedDocuments > 0);
+        } finally {
+            Iterator<VerificationHost> peerIt = this.host.getInProcessHostMap().values().iterator();
+            while (peerIt.hasNext()) {
+                VerificationHost peerNode = peerIt.next();
+                this.host.stopHost(peerNode);
+            }
+        }
     }
 
     @Test