+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.5.4-SNAPSHOT
 
+* Updates to migration task service to assert zero documentOwner mismatches 
+  while migrating services that don't use ServiceOption ON_DEMAND_LOAD.
+  This avoids cases when some stateful services fail to migrate due because
+  of pending synchronization.
+
 ## 1.5.3
 
 * Addition of new "indexed metadata" document indexing and query options.